Update docs to mention CUDA support for Future
To reflect the many changes introduced recently.

In my mind, CUDAFuture should be considered a "private" subclass, which in practice should always be returned as a downcast pointer to an ivalue::Future. Hence, we should document the CUDA behavior in the superclass, even if it's CUDA-agnostic, since that's the interface the users will see also for CUDA-enabled futures.

Differential Revision: [D25757474](https://our.internmc.facebook.com/intern/diff/D25757474/)

ghstack-source-id: 119291120
Pull Request resolved: https://github.com/pytorch/pytorch/pull/50048